Make a directory for the data parser
